Acceleratorclass refactoring towards optics measurements working in all machines
